Student

+++ next-app/auth.config.ts
@@ -2,6 +2,10 @@ import CredentialsProvider from "next-auth/providers/credentials";
 import Google from "next-auth/providers/google";
 
 import {
+  getStudentByIdData,
+  getStudentByIdentifierlData,
+} from "@/data/student-data";
+import {
   deleteTwoFactorComfirmationByIdData,
   getTwoFactorComfirmationByUserIdData,
 } from "@/data/two-factor-comfirmation.data";
@@ -15,8 +19,11 @@ import {
   getVerificationTokenByIdentifierData,
 } from "@/data/verification-token-data";
 import { verifyPassword } from "@/lib/hash-password";
-import { loginUserFormSchema } from "@/shema-zod/auth.shema";
-import { UserRole } from "@prisma/client";
+import {
+  loginStudentFormSchema,
+  loginUserFormSchema,
+} from "@/shema-zod/auth.shema";
+import { Student, User, UserRole } from "@prisma/client";
 import type { NextAuthConfig } from "next-auth";
 import "next-auth/jwt";
 
@@ -24,8 +31,9 @@ declare module "next-auth" {
   interface User {
     /** The user role. */
     role: UserRole;
-    isTwoFactorEnabled: boolean;
+    isTwoFactorEnabled?: boolean;
     isOAuth?: boolean;
+    identifier?: string;
   }
 }
 
@@ -34,8 +42,9 @@ declare module "next-auth/jwt" {
   interface JWT {
     /** The user role. */
     role: UserRole;
-    isTwoFactorEnabled: boolean;
-    isOAuth: boolean;
+    isTwoFactorEnabled?: boolean;
+    isOAuth?: boolean;
+    identifier?: string;
   }
 }
 
@@ -47,9 +56,14 @@ export default {
   providers: [
     CredentialsProvider({
       async authorize(credentials) {
-        const isCredentialsValide = loginUserFormSchema.safeParse(credentials);
-        if (isCredentialsValide.success) {
-          const { email, password } = isCredentialsValide.data;
+        const isCredentialsAdminValide =
+          loginUserFormSchema.safeParse(credentials); // Vérification Admin
+        const isCredentialsStudentValide =
+          loginStudentFormSchema.safeParse(credentials); // Vérification Student
+
+        // Vérification Admin
+        if (isCredentialsAdminValide.success) {
+          const { email, password } = isCredentialsAdminValide.data;
 
           const user = await getUserByEmailData(email);
           if (!user || !user.password) return null;
@@ -60,6 +74,23 @@ export default {
           }
         }
 
+        // Vérification student
+        if (isCredentialsStudentValide.success) {
+          const { identifier, password } = isCredentialsStudentValide.data;
+
+          const student = await getStudentByIdentifierlData(identifier);
+          if (!student || !student.password) return null;
+
+          const passwordMatch = await verifyPassword(
+            student.password,
+            password
+          );
+          if (passwordMatch) {
+            return student;
+          }
+        }
+
+        // Si rien n'est bon on retourne null
         return null;
       },
     }),
@@ -69,40 +100,47 @@ export default {
     async signIn({ user, account }) {
       // Allow OAuth without email verification
       if (account?.provider !== "credentials") {
-        // TODO: Vérifier si il y'a un token d'ajout à un groupe
-        // if (account?.provider === "google") {
-        //   const googleUser = await getUserByEmailData(user?.email as string);
-        //   if (googleUser) return true;
-        // }
-
         return true;
       }
 
       if (!user) return false;
 
-      const existingUser = await getUserByIdData(user.id as string);
+      // Vérification si role est ADMIN
+      if (user.role === "ADMIN") {
+        const existingUser = await getUserByIdData(user.id as string);
 
-      // Prevent sign in without email verification
-      if (!existingUser?.email || !existingUser?.emailVerified) return false;
+        // Prevent sign in without email verification
+        if (!existingUser?.email || !existingUser?.emailVerified) return false;
 
-      const existingVerification = await getVerificationTokenByIdentifierData(
-        existingUser.email
-      );
+        const existingVerification = await getVerificationTokenByIdentifierData(
+          existingUser.email
+        );
 
-      // Deletion of the token if the email has been verified
-      if (existingVerification)
-        await deleteVerificationTokenByIdData(existingVerification.id);
+        // Deletion of the token if the email has been verified
+        if (existingVerification)
+          await deleteVerificationTokenByIdData(existingVerification.id);
 
-      if (existingUser.isTwoFactorEnabled) {
-        const twoFactorComfirmation =
-          await getTwoFactorComfirmationByUserIdData(existingUser.id);
+        if (existingUser.isTwoFactorEnabled) {
+          const twoFactorComfirmation =
+            await getTwoFactorComfirmationByUserIdData(existingUser.id);
 
-        if (!twoFactorComfirmation) return false;
+          if (!twoFactorComfirmation) return false;
 
-        await deleteTwoFactorComfirmationByIdData(twoFactorComfirmation.id);
+          await deleteTwoFactorComfirmationByIdData(twoFactorComfirmation.id);
+        }
+
+        return true;
       }
 
-      return true;
+      // Vérification si role est STUDENT
+      if (user.role === "STUDENT") {
+        const existingUser = await getStudentByIdData(user.id as string);
+        if (!existingUser) return false;
+
+        return true;
+      }
+
+      return false;
     },
     async jwt({ token, account, trigger }) {
       if (!token.sub) return token;
@@ -110,13 +148,25 @@ export default {
       if (trigger === "signIn") token.isOAuth = !!account?.access_token;
 
       if (trigger === "signIn" || trigger === "update") {
-        const existingUser = await getUserByIdData(token.sub);
-        if (!existingUser) return token;
-        token.name = existingUser.name;
-        token.email = existingUser.email;
-        token.picture = existingUser.image;
-        token.role = existingUser.role;
-        token.isTwoFactorEnabled = existingUser.isTwoFactorEnabled;
+        // let existingData: User | Student | null;
+        const existingUserData: User | null = await getUserByIdData(token.sub);
+        const existingStudentData: Student | null = await getStudentByIdData(
+          token.sub
+        );
+        // Récupère les données depuis le modèle approprié en fonction du type des données
+        if (existingStudentData) {
+          token.identifier = existingStudentData?.identifier;
+          token.role = existingStudentData.role;
+          token.name = existingStudentData.name;
+          token.picture = existingStudentData.image;
+        } else if (existingUserData) {
+          token.email = existingUserData?.email || "";
+          token.isTwoFactorEnabled =
+            existingUserData?.isTwoFactorEnabled || false;
+          token.role = existingUserData.role;
+          token.name = existingUserData.name;
+          token.picture = existingUserData.image;
+        }
       }
 
       return token;
@@ -126,6 +176,7 @@ export default {
         ...session,
         user: {
           ...session.user,
+          identifier: token.identifier,
           role: token.role,
           isTwoFactorEnabled: token.isTwoFactorEnabled,
           isOAuth: token.isOAuth,
